move predicted kalman drops into EventPerigee class
